added several changes:
- plugins can now load their own locales. The PluginManager will take care of it at registration time
and use the Locales class to store those new strings. When a locale is requested via the Locales::getLocale
method, the strings loaded from the locale under locale/ will be merged with the strings that were loaded
from the plugins. If the current language is not available from the plugin, en_UK will be used.
- added a new field to the plog_custom_fields_definition table, so that fields can be marked as hidden. Hidden
fields will still be fetched but they won't be shown when adding a new post or editing a post that already exists.
It is necessary to rerun the wizard after these changes.